% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingHelperFunctions.R
\name{axis_break}
\alias{axis_break}
\title{places a break marker at the position "breakpos" in user coordinates on the axis nominated}
\usage{
axis_break(
  axis = 1,
  breakpos,
  bgcol = "white",
  breakcol = "black",
  style = c("slash"),
  brw = 0.02,
  Lwd = 1
)
}
\arguments{
\item{axis}{numeric: which axis the break will be on}

\item{breakpos}{numeric: where the break will be}

\item{bgcol}{character string: the background color; white by default}

\item{breakcol}{character string: the color of the axis break; black by default}

\item{style}{list of strings: whether the break is a slash or a zigzag}

\item{brw}{numeric: the break width; 0.02 by default}

\item{Lwd}{numeric: line width for the plot; 1 by default}
}
\value{
The current plot with zigzag axis breaks at the desired place, to better fit data in a visualization
}
\description{
places a break marker at the position "breakpos" in user coordinates on the axis nominated
}
