% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunctions.R
\name{categoryMahalanobisDistances}
\alias{categoryMahalanobisDistances}
\title{Calculate category mahalanobis distances for pipeline data}
\usage{
categoryMahalanobisDistances(
  Level5,
  FeatureList,
  CategoryName,
  coverVariance,
  minObjects,
  SType = "vehicle control",
  mongoUrl = "",
  use_db = TRUE,
  json_collection_path = ""
)
}
\arguments{
\item{Level5}{table: A table of well data at "level 5" in the pipeline}

\item{FeatureList}{matrix: The features for each category}

\item{CategoryName}{character string: The category whose variances are being compared}

\item{coverVariance}{numeric: The known variance of the well data}

\item{minObjects}{numeric: The minimum number of expected objects}

\item{SType}{character string: Defines which sample type will be used for data normalization; "vehicle control" by default}

\item{mongoUrl}{character string: the database where the collections are be stored and the required credentials, generated by the mongoURL function}

\item{use_db}{boolean: Determines whether mongoDB will be used or not; default is TRUE}

\item{json_collection_path}{character: Full file path to where JSON collections will be stored}
}
\value{
VarianceExplainedList a list of variances with category Mahalanobis distances calculated
}
\description{
Calculate category mahalanobis distances for pipeline data
}
