% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{cellViability_plots}
\alias{cellViability_plots}
\title{This function will generate a plot for each chem_id in the HTPP dataset (for each cell type if applicable).  Creates a plot of cell viability data for each chemical id in the dataset in the location specfied by filepath.}
\usage{
cellViability_plots(file_path, study_name, mongoUrl, refChems = TRUE)
}
\arguments{
\item{file_path}{character string: path to directory where HTPP plots will be populated}

\item{study_name}{character string: name of study to be used for plot titles}

\item{mongoUrl}{character string: the database where the collections are be stored and the required credentials, generated by the mongoURL function}

\item{refChems}{boolean: only make viability plots for reference chemicals instead of all chemicals; default in TRUE}
}
\description{
This function will generate a plot for each chem_id in the HTPP dataset (for each cell type if applicable).  Creates a plot of cell viability data for each chemical id in the dataset in the location specfied by filepath.
}
