% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingHelperFunctions.R
\name{Cluster.by.Channel}
\alias{Cluster.by.Channel}
\title{Orders the columns to be a heatmap axis}
\usage{
Cluster.by.Channel(Data)
}
\arguments{
\item{Data}{table to be reordered}
}
\value{
newOrder
}
\description{
Orders the columns to be a heatmap axis
}
