% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingHelperFunctions.R
\name{ColumnColors}
\alias{ColumnColors}
\title{Assigns a color to each column of data by name}
\usage{
ColumnColors(Data, dataform = "matrix")
}
\arguments{
\item{Data}{dataframe: the data being assigned color}

\item{dataform}{character string: whether the data is a "matrix" or a "vector"}
}
\value{
array of colors matching each of the columns in the data, to distinguish them in plots
}
\description{
Assigns a color to each column of data by name
}
