% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingHelperFunctions.R
\name{concRespPlot_JN}
\alias{concRespPlot_JN}
\title{Concentration Response Plot}
\usage{
concRespPlot_JN(row, ymin = NULL, ymax = NULL)
}
\arguments{
\item{row}{vector: Row containing response data to be plotted}

\item{ymin}{numeric: y axis minimum}

\item{ymax}{numeric: y axis maximum}
}
\value{
a log concentration-response plot
}
\description{
Concentration Response Plot
}
\examples{
conc <- list(.03, .1, .3, 1, 3, 10, 30, 100)
resp <- list(0, .2, .1, .4, .7, .9, .6, 1.2)
row <- list(conc = conc,
            resp = resp,
            bmed = 0,
            cutoff = 0.25,
            onesd = 0.125,
            name = "some chemical",
            assay = "some assay")
res <- tcplfit2::concRespCore(row, conthits = TRUE)
res <- dplyr::mutate(.data = res,
                     la = NA,
                     q = NA,
                     stype = "test sample",
                     endpoint = "test")
Subset <- dplyr::mutate(.data = res,
                        acc = ifelse(is.na(acc), bmd, acc))
Subset2 <- dplyr::filter(.data = Subset,
                         stype \%in\% c("test sample") & bmd < 100 & hitcall > 0.90)
concRespPlot_JN(Subset2,ymin=-0.5,ymax=2)

}
