% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingHelperFunctions.R
\name{correctChemName}
\alias{correctChemName}
\title{Replaces colons in chemical names with '-' so they will be read in correctly}
\usage{
correctChemName(String)
}
\arguments{
\item{String}{character string: chem names to correct}
}
\value{
chem_names that fit the correct pattern
}
\description{
Replaces colons in chemical names with '-' so they will be read in correctly
}
\examples{
correctChemName("test:chemical")

}
