% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{curvePlots_htppGlobalMah}
\alias{curvePlots_htppGlobalMah}
\title{Plots global mahalanobis distances from htpp_tcpl collection, writes plots of global data to the location specified by file_path}
\usage{
curvePlots_htppGlobalMah(file_path, study_name, mongoUrl)
}
\arguments{
\item{file_path}{character: file path to where global mah plots will be created}

\item{study_name}{character string: name of study to be used for plot titles, should follow a similar naming convention used in other functions such as "viability_controlPlot_htppWell.R"}

\item{mongoUrl}{haracter string: URL to connect to MongoDB for HTPP dataset; can be created using the mongoURL function in htpp.pl}
}
\description{
Plots global mahalanobis distances from htpp_tcpl collection, writes plots of global data to the location specified by file_path
}
