% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunctions.R
\name{CVanalysis}
\alias{CVanalysis}
\title{Reformats the data into a Mongo collection, normalizes it based on the solvent control and finds the percent responder cells.  Inputs the cv_well and cv_image_metadata collections.}
\usage{
CVanalysis(
  InputPath,
  PlateID,
  SType = "vehicle control",
  mongoUrl,
  minNucleiArea = 30,
  maxNucleiArea = 1000,
  minRoundness = 0.5
)
}
\arguments{
\item{InputPath}{character string: the input path to the Harmony file}

\item{PlateID}{character string: the plate_id value}

\item{SType}{character string: Defines which sample type will be used for data normalization; "vehicle control" by default}

\item{mongoUrl}{character string: The MongoDB host, user, password and database}

\item{minNucleiArea}{numeric: The minimum area for something flagged as nucleus for QC}

\item{maxNucleiArea}{numeric: The maximum area for something flagged as nucleus for QC}

\item{minRoundness}{numeric: The minimum cell roundness for something to be recognized as a cell}
}
\description{
Reformats the data into a Mongo collection, normalizes it based on the solvent control and finds the percent responder cells.  Inputs the cv_well and cv_image_metadata collections.
}
