% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{Euclidean_norm_vec}
\alias{Euclidean_norm_vec}
\title{Calculate the Euclidean norm of a vector}
\usage{
Euclidean_norm_vec(vect)
}
\arguments{
\item{vect}{Numeric vector: the vector you will take the euclidean norm of}
}
\value{
Numeric: the Euclidean norm of the vector
}
\description{
Calculate the Euclidean norm of a vector
}
\examples{
Euclidean_norm_vec(c(1,5,3,4,12))
}
