% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingHelperFunctions.R
\name{findDataCols}
\alias{findDataCols}
\title{finds columns by name, if inv is true, only finds those columns that exist in the data}
\usage{
findDataCols(Table, inv = F, names = F)
}
\arguments{
\item{Table}{A table of cell-painting data}

\item{inv}{A boolean, if it's TRUE, findDataCols only finds those columns that exist in the data}

\item{names}{A boolean, if it's TRUE, return only the column names, if it's FALSe, return the columns}
}
\value{
Either the column names, or the columns in Table, based on whether names is TRUE
}
\description{
finds columns by name, if inv is true, only finds those columns that exist in the data
}
