% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCollections.R
\name{generate_htppCatMah}
\alias{generate_htppCatMah}
\title{Create htpp collection htpp_cat_mah for category Mahalanobis distances}
\usage{
generate_htppCatMah(
  coverVariance,
  minObjects,
  mongoUrl,
  varianceExplainedPath,
  nThreads = 1,
  rerun = FALSE
)
}
\arguments{
\item{coverVariance}{numeric: The value of variance explained used to determine the number of eigen features used in analysis}

\item{minObjects}{numeric: The minimum number of objects used to filter the dataset for analysis}

\item{mongoUrl}{character string: URL to connect to MongoDB for HTPP dataset; can be created using the mongoURL function in htpp.pl}

\item{varianceExplainedPath}{character string: the path where the function will write variance explained metadata}

\item{nThreads}{numeric:  the number of threads to use for processing; default is 1}

\item{rerun}{boolean: rerun = TRUE will drop existing htpp_cat_mah collection and reinsert; have FALSE by default}
}
\description{
Create htpp collection htpp_cat_mah for category Mahalanobis distances
}
