% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCollections.R
\name{generate_htppFeature_htppCategory}
\alias{generate_htppFeature_htppCategory}
\title{Inserts feature and category data into mongo collection htpp_feature and htpp_category}
\usage{
generate_htppFeature_htppCategory(
  inputPath,
  PlateID,
  Cell_Type,
  mongoUrl = "",
  file_path = "",
  rerun = FALSE,
  use_db = TRUE,
  json_collection_path = ""
)
}
\arguments{
\item{inputPath}{character string: Can either be a truncated path, or a full path to a HTPP data file. If it is truncated, the function will rebuild a full path using file_path}

\item{PlateID}{character string: The PlateID for the plate being analyzed}

\item{Cell_Type}{character: The cell type associated with the specific feature and category set used for a set of HTPP plates; will be saved to collections}

\item{mongoUrl}{character string: The database where the collections will be stored}

\item{file_path}{character string: The path to where the input file is located}

\item{rerun}{boolean:  Whether to delete and reinsert into both collections; false by default}

\item{use_db}{boolean. Determines whether mongoDB will be used or not; default is TRUE}

\item{json_collection_path}{character. Full file path to where JSON collections will be stored}
}
\description{
Inserts feature and category data into mongo collection htpp_feature and htpp_category
}
