% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCollections.R
\name{generate_htppNullChems}
\alias{generate_htppNullChems}
\title{Create plots for signal strength and add NULL chemicals to htpp_well_norm}
\usage{
generate_htppNullChems(
  n_lowest_conc = 2,
  n_cv_active_conc = 6,
  rel_cellCount = 50,
  plot_file_path,
  study_name,
  mongoUrl,
  ConcList = c(100, 10, 1, 0.1, 0.01, 0.001, 1e-04, 1e-05),
  rerun = FALSE
)
}
\arguments{
\item{n_lowest_conc}{integer: The number of the lowest concentrations in a concentration series for modeling Null chemical data; Default is 2 (i.e., dose_level 1 and 2)}

\item{n_cv_active_conc}{integer: The number of cell viability active concentrations to be excluded; default is 6 (i.e., exclude chemicals where dose_level >= 6 are cell viability actives)}

\item{rel_cellCount}{integer: The relative cell count threshold for excluding well data for Null chemical sampling; default is 50 (i.e., exclude wells with rel_cell_count < 50)}

\item{plot_file_path}{character string: file path where signal strength plots will be created}

\item{study_name}{character string: the name of the study}

\item{mongoUrl}{character string: URL to connect to MongoDB for HTPP dataset; can be created using the mongoURL function in htpp.pl}

\item{ConcList}{numeric vector: vector of 8 test concentrations to be used for the NULL chemicals. c(100, 10, 1, 0.1, 0.01, 0.001, 0.0001, 0.00001) by default.}

\item{rerun}{boolean: rerun = TRUE will drop existing cv_well collection and reinsert; FALSE by default}
}
\description{
Create plots for signal strength and add NULL chemicals to htpp_well_norm
}
