% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCollections.R
\name{generate_htppPac_globalMah}
\alias{generate_htppPac_globalMah}
\title{generate htpp_pac from htpp_bmc and add global mahalanobis distance records to htpp_pac}
\usage{
generate_htppPac_globalMah(mongoUrl, hit_n_conc = 4, rerun = FALSE)
}
\arguments{
\item{mongoUrl}{character string: URL to connect to MongoDB for HTPP dataset; can be created using the mongoURL function in htpp.pl}

\item{hit_n_conc}{numeric: Number of test concentrations needed during curve fitting to determine if a PAC is a hit; default is 4}

\item{rerun}{Boolean: TRUE will drop existing entries in htpp_bmc for approach = "global" and endpoint = "global", and reinsert; default is FALSE}
}
\description{
generate htpp_pac from htpp_bmc and add global mahalanobis distance records to htpp_pac
}
