% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCollections.R
\name{generate_htppProfile}
\alias{generate_htppProfile}
\title{create htpp_profile collection}
\usage{
generate_htppProfile(n_cells, relative_cellCount, mongoUrl, rerun = FALSE)
}
\arguments{
\item{n_cells}{numeric: Minimum threshold for the number of cells to keep for filtering}

\item{relative_cellCount}{numeric: Minimum threshold of the count of relative number of cells to use for filtering}

\item{mongoUrl}{character string: URL to connect to MongoDB for HTPP dataset; can be created using the mongoURL function in htpp.pl}

\item{rerun}{boolean: rerun = TRUE will drop existing collection and reinsert; FALSE by default}
}
\description{
create htpp_profile collection
}
