% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCollections.R
\name{generate_htppStudy}
\alias{generate_htppStudy}
\title{Adding study-level metadata to htpp_study collection}
\usage{
generate_htppStudy(
  input,
  mongoUrl = "",
  rerun = FALSE,
  use_db = TRUE,
  json_collection_path = ""
)
}
\arguments{
\item{input}{data.frame/data.table: A data frame with study information; should include the following columns: "cell_type", "stype", "exposure_duration", "seeding_density"}

\item{mongoUrl}{character string: A mongoUrl with credentials to access the database}

\item{rerun}{boolean: Whether you want to clear the mongo database as you go and refill it; false by default}

\item{use_db}{boolean: Determines whether mongoDB will be used or not; default is TRUE}

\item{json_collection_path}{character: Full file path to where JSON collections will be stored}
}
\description{
Adding study-level metadata to htpp_study collection
}
