% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCollections.R
\name{generate_htppWell}
\alias{generate_htppWell}
\title{Creates htppWell collections (htpp_well_raw, htpp_well, htpp_image_metadata)}
\usage{
generate_htppWell(
  file_path,
  mongoUrl = "",
  Cell_Type,
  CellArea.Limit,
  NucleiArea.Limit,
  SType = "vehicle control",
  n_max = 2000 * 384,
  rerun = FALSE,
  replace = FALSE,
  use_db = TRUE,
  json_collection_path = ""
)
}
\arguments{
\item{file_path}{character string: file path to the top level directory of Harmony files for an HTPP dataset (i.e., the directory above plate-level directories)}

\item{mongoUrl}{character string: The URL of the mongo database holding the collection, with user credentials to access it}

\item{Cell_Type}{character string or list of strings: the cell type or types being used}

\item{CellArea.Limit}{dictionary: A dictionary of cells and their corresponding cell area limits, of the form c("celltype" = list(c(lower,upper)))}

\item{NucleiArea.Limit}{dictionary: A dictionary of cells and their corresponding nuclei area limits, of the form c("celltype" = list(c(lower,upper)))}

\item{SType}{character string: The sample type used for normalization. Set to "vehicle control" by default}

\item{n_max}{numeric: The maximum dimensions of the table}

\item{rerun}{boolean: Whether to drop and replace the collections in the dataframe before inserting}

\item{replace}{boolean: Whether you want to replace existing records in the mongo database; false by default}

\item{use_db}{boolean: Determines whether mongoDB will be used or not; default is TRUE}

\item{json_collection_path}{character: Full file path to where JSON collections will be stored}
}
\description{
Creates htppWell collections (htpp_well_raw, htpp_well, htpp_image_metadata)
}
