% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCollections.R
\name{generate_htppWellNorm}
\alias{generate_htppWellNorm}
\title{Create http_well_norm, a collection of normalized well data for all plate groups}
\usage{
generate_htppWellNorm(mongoUrl, rerun = FALSE)
}
\arguments{
\item{mongoUrl}{character string: URL to connect to MongoDB for HTPP dataset; can be created using the mongoURL function in htpp.pl}

\item{rerun}{boolean: rerun = TRUE will drop existing collection and reinsert; have FALSE by default}
}
\description{
Create http_well_norm, a collection of normalized well data for all plate groups
}
