% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCollections.R
\name{generate_htppWellTrt_htppChem}
\alias{generate_htppWellTrt_htppChem}
\title{Adding chemical id and sample data to htpp_well_trt, htpp_chem}
\usage{
generate_htppWellTrt_htppChem(
  SampleKey,
  mongoUrl,
  rerun = FALSE,
  replace = TRUE
)
}
\arguments{
\item{SampleKey}{dataframe: The sample key with chemical info and metadata}

\item{mongoUrl}{characters string A mongoUrl with credentials to access the database}

\item{rerun}{boolean: Whether you want to clear the mongo database as you go and refill it; false by default}

\item{replace}{boolean: Whether you want to replace existing records in the mongo database; false by default}
}
\description{
Adding chemical id and sample data to htpp_well_trt, htpp_chem
}
