% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingHelperFunctions.R
\name{GOLetter}
\alias{GOLetter}
\title{gives one letter abbreviations to attributes}
\usage{
GOLetter(Vector)
}
\arguments{
\item{Vector}{a vector of cell attributes}
}
\value{
vector of single-letter abbreviations of the attributes listed, to distinguish points on a plot
}
\description{
gives one letter abbreviations to attributes
}
\examples{
GOLetter(c("test"))

}
