% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{maxJN}
\alias{maxJN}
\title{calculate the maximum, but return NA is no number is present (pmax doesn't work with summarise)}
\usage{
maxJN(x)
}
\arguments{
\item{x}{a vector one wants the maximum of}
}
\value{
The maximum, or NA if there are no nonzero numbers in X
}
\description{
calculate the maximum, but return NA is no number is present (pmax doesn't work with summarise)
}
\examples{
maxJN(c(4,1,7,10,9))
maxJN(c(NA, NA, NA, NA, NA, NA, NA, NA, 0))
}
