% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{minJN}
\alias{minJN}
\title{calculate the minimum, but return NA is no number is present (pmin doesn't work with summarise)}
\usage{
minJN(x)
}
\arguments{
\item{x}{a vector one wants the minimum of}
}
\value{
The minimum, or NA if there are no nonzero numbers in X
}
\description{
calculate the minimum, but return NA is no number is present (pmin doesn't work with summarise)
}
\examples{
minJN(c(4,1,7,10,9))
minJN(c(NA, NA, NA, NA, NA, NA, NA, NA, 0))
}
