% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{mongoQuery}
\alias{mongoQuery}
\title{Build a mongo query from an arbitrary set of params or a list}
\usage{
mongoQuery(...)
}
\arguments{
\item{...}{(any) = Any set of named parameters OR a single named list of all arguments}
}
\value{
JSON query ready to pass to mongolite functions
}
\description{
Build a mongo query from an arbitrary set of params or a list
}
\examples{
mongoQuery(approach = "global", endpoint = "global")
}
