% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{outerFences}
\alias{outerFences}
\title{Tukey Outer fence function}
\usage{
outerFences(input_vector, iqr.factor = 3)
}
\arguments{
\item{input_vector}{vector: the data whose outer fences you want to calculate}

\item{iqr.factor}{numeric: the interquartile range factor.  3 by default.}
}
\value{
upper and lower fences
}
\description{
Tukey Outer fence function
}
\examples{
outerFences(input_vector= c(113.86844, 108.47126,
125.22345, 115.17092, 123.61978, 112.45098, 128.88594,
100.98654, 103.95449, 109.41060, 114.59485, 107.66969,
101.75302, 100.46038, 103.86191, 110.25791,
113.22980, 101.13067, 112.93010, 105.88312, 98.46702,
92.14626, 97.33307, 117.61402, 110.81441,
106.10610, 110.93701, 115.87897, 116.07416, 104.51375,
106.85793, 117.94644, 111.48693, 122.34972,
103.06462, 127.57024, 120.70566, 98.62746, 110.22989,
150.98643), iqr.factor=3)

}
