% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingHelperFunctions.R
\name{PchShape}
\alias{PchShape}
\title{PchShape}
\usage{
PchShape(Parameter.Name)
}
\arguments{
\item{Parameter.Name}{character string: the parameter names being assigned graph shapes}
}
\value{
vector of 2-digit base r plots codes for single-letter abbreviations of the attributes listed, to distinguish points on a plot
}
\description{
PchShape
}
\examples{
PchShape("AGP")

}
