% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingHelperFunctions.R
\name{plotCurves}
\alias{plotCurves}
\title{Plots concentration-response data to visualize Benchmark Dose (BMD)}
\usage{
plotCurves(
  Subset,
  xLim = NULL,
  TestedRange = NULL,
  plotDatapoints = F,
  plotBMC = F,
  plotDoserange = F,
  plotNoiseband = T,
  Lwd = 1,
  cexAxis = 1,
  yLim = c(-5.5, 100),
  yTicks = NULL,
  yAxisSteps = 25
)
}
\arguments{
\item{Subset}{dataframe: the subset of the data to be plotted, for instance subset by a given chemical}

\item{xLim}{numeric: x limit of the plot; null by default.}

\item{TestedRange}{the dose range tested; null by default.}

\item{plotDatapoints}{boolean: whether the individual datapoints should be visualized on the plot}

\item{plotBMC}{boolean: whether the BMC should be highlighted on the plot; false by default}

\item{plotDoserange}{boolean: whether the doserange should be plotted; false by default}

\item{plotNoiseband}{boolean: whether the noise band should be plotted; true by default}

\item{Lwd}{numeric: line width on the plot; 1 by default}

\item{cexAxis}{numeric: scale of the axis labels; 1 by default}

\item{yLim}{numeric: y limit of the plot; c(-5.5, 100) by default}

\item{yTicks}{numeric: the placement of y axis ticks; null by default.}

\item{yAxisSteps}{the frequency of y axis ticks, if yTicks is undefined}
}
\value{
hill plot of the data
}
\description{
Plots concentration-response data to visualize Benchmark Dose (BMD)
}
