% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_htpp_sampleKey.R
\name{validate_htpp_sampleKey}
\alias{validate_htpp_sampleKey}
\title{Reformatting the sample key into a more machine-readable form for later processing and checking it for errors}
\usage{
validate_htpp_sampleKey(
  SampleKey,
  skipped_tests = c(),
  max_dose_level = 8,
  dataFrame = FALSE
)
}
\arguments{
\item{SampleKey}{File name, path to file, of the sample key file (in csv format) being used}

\item{skipped_tests}{The names of QC tests you want to skip}

\item{max_dose_level}{The maximum dose level used}

\item{dataFrame}{Boolean, if TRUE, will treat SampleTable as data.frame/data.table input and not read in file; default is FALSE}
}
\value{
A reformatted sample key if there were no errors, or a list of errors if there were
}
\description{
Reformatting the sample key into a more machine-readable form for later processing and checking it for errors
}
\examples{
sample_key <- data.table::fread(
file = system.file("extdata", "example_sampleKey.csv", package = "htpp.pl"),
sep = ",")

validated_sample_key <- validate_htpp_sampleKey(SampleKey = sample_key,
max_dose_level = 8, skipped_tests = "QCV_0", dataFrame = TRUE)

}
