% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{viability_controlPlot_htppWell}
\alias{viability_controlPlot_htppWell}
\title{Create plots from htpp_well and write them to the location specified in file_path}
\usage{
viability_controlPlot_htppWell(
  file_path,
  vehicle_control,
  viability_positive_control,
  study_name,
  mongoUrl = "",
  use_db = T,
  json_collection_path = ""
)
}
\arguments{
\item{file_path}{character string: The path to where the plots will go.}

\item{vehicle_control}{character string: chem_id for vehicle control samples}

\item{viability_positive_control}{character string: chem_id for viability positive control samples}

\item{study_name}{character string: name of study to be used for plot titles}

\item{mongoUrl}{character string: The URL of the mongo database holding the collection, with user credentials to access it}

\item{use_db}{boolean: Determines whether mongoDB will be used or not; default is TRUE}

\item{json_collection_path}{character: Full file path to where JSON collections will be stored}
}
\description{
Create plots from htpp_well and write them to the location specified in file_path
}
