% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{Well3Digit}
\alias{Well3Digit}
\title{Convert 2 digit values to 3 digit in a vector by adding a preceding 0}
\usage{
Well3Digit(Vector)
}
\arguments{
\item{Vector}{The vector being modified}
}
\value{
The vector with all 3 digit values
}
\description{
Convert 2 digit values to 3 digit in a vector by adding a preceding 0
}
\examples{
Well3Digit(c("100","10","30"))
}
