% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellViability.R
\name{generate_cvBMC}
\alias{generate_cvBMC}
\title{Creates and populates cell viability bmc (cv_bmc) collection in mongo}
\usage{
generate_cvBMC(mongoUrl, rerun = FALSE)
}
\arguments{
\item{mongoUrl}{character string: URL to connect to MongoDB for HTPP dataset; can be created using the mongoURL function in htpp.pl}

\item{rerun}{TRUE will drop existing collection and reinsert; FALSE by default}
}
\description{
Creates and populates cell viability bmc (cv_bmc) collection in mongo
}
