% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellViability.R
\name{generate_cvTcpl}
\alias{generate_cvTcpl}
\title{Creates cell viability collection cv_tcpl based on well and chem data}
\usage{
generate_cvTcpl(cell_viability, mongoUrl, rerun = FALSE)
}
\arguments{
\item{cell_viability}{boolean: if cell_viability = TRUE, the CVData_all object in line 218 should pull out data from the cv_well collection, otherwise it will pull data from the htpp_well collection}

\item{mongoUrl}{character string: URL to connect to MongoDB for HTPP dataset; can be created using the mongoURL function in htpp.pl}

\item{rerun}{boolean: rerun = TRUE will drop existing cv_tcpl collection and reinsert; FALSE by default}
}
\description{
Creates cell viability collection cv_tcpl based on well and chem data
}
