% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCollections.R
\name{generate_htppFeature_htppCategory}
\alias{generate_htppFeature_htppCategory}
\title{Inserts feature and category data into mongo collection htpp_feature and htpp_category}
\usage{
generate_htppFeature_htppCategory(
  inputPath,
  PlateID,
  mongoUrl,
  file_path = "",
  rerun = FALSE
)
}
\arguments{
\item{inputPath}{character string: Can either be a truncated path, or a full path to a HTPP data file. If it is truncated, the function will rebuild a full path using file_path}

\item{PlateID}{character string: The PlateID for the plate being analyzed}

\item{mongoUrl}{character string: The database where the collections will be stored}

\item{file_path}{character string: The path to where the input file is located}

\item{rerun}{boolean:  Whether to delete and reinsert into both collections; false by default}
}
\description{
Inserts feature and category data into mongo collection htpp_feature and htpp_category
}
