% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCollections.R
\name{generate_htppGlobalMah}
\alias{generate_htppGlobalMah}
\title{Calculate, plot and record global Mahalanobis distances from mongo data}
\usage{
generate_htppGlobalMah(
  coverVariance,
  minObjects,
  plot_file_path,
  study_name,
  mongoUrl,
  rerun = FALSE
)
}
\arguments{
\item{coverVariance}{numeric: The value of variance explained used to determine the number of eigen features used in analysis}

\item{minObjects}{numeric: The minimum number of objects used to filter the dataset for analysis}

\item{plot_file_path}{character string: file path where variance explained plots will be created}

\item{study_name}{character string: the name of the experiment used to title the plots}

\item{mongoUrl}{URL to connect to MongoDB for HTPP dataset; can be created using the mongoURL function in htpp.pl}

\item{rerun}{boolean: rerun = TRUE will drop existing cv_well collection and reinsert; FALSE by default}
}
\description{
Calculate, plot and record global Mahalanobis distances from mongo data
}
