% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunctions.R
\name{globalMahalanobisDistances}
\alias{globalMahalanobisDistances}
\title{Calculate global Mahalanobis distances for a table}
\usage{
globalMahalanobisDistances(
  Table1,
  coverVariance,
  minObjects,
  SType = "vehicle control",
  url = "",
  use_db = TRUE,
  json_collection_path = ""
)
}
\arguments{
\item{Table1}{table: A table of well data}

\item{coverVariance}{numeric: The known variance of the well data}

\item{minObjects}{numeric: Minimum number of cells for plate to pass QC filter}

\item{SType}{character string: Defines which sample type will be used for data normalization; "vehicle control" by default}

\item{url}{character string: The MongoDB host, user, password and database}

\item{use_db}{boolean: Determines whether mongoDB will be used or not; default is TRUE}

\item{json_collection_path}{character: Full file path to where JSON collections will be stored}
}
\value{
A list of cumulative proportion, rotation matrix and inverse covariance
}
\description{
Calculate global Mahalanobis distances for a table
}
