% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunctions.R
\name{Raw2Level4}
\alias{Raw2Level4}
\title{Summarize, clean and format the raw data, and store it as a Mongo collection.  Run basic QC before other steps}
\usage{
Raw2Level4(
  InputPath,
  PlateID,
  mongoUrl,
  Cell_Type = NULL,
  CellArea.Limit = NULL,
  NucleiArea.Limit = NULL,
  n_max = 2000 * 384,
  SType = "vehicle control"
)
}
\arguments{
\item{InputPath}{character string: Path to the Harmony file}

\item{PlateID}{character string: the plate_id for the well plate}

\item{mongoUrl}{character string: The URL of the mongo database holding the collection, with user credentials to access it}

\item{Cell_Type}{character string: the cell line used; null by default.}

\item{CellArea.Limit}{list: The size range for cells examined.  If the Cell_Type is "U-2 OS", "MCF7", "A549", "ARPE-19", "HepG2" or "HTB-9", this will autofill with their ranges.  Otherwise list(c(0, 99999999)) by default.}

\item{NucleiArea.Limit}{list: The nucleus area range for cells examined.  If the Cell_Type is "U-2 OS", "MCF7", "A549", "ARPE-19", "HepG2" or "HTB-9", this will autofill with their ranges.  Otherwise list(c(0,99999999)) by default.}

\item{n_max}{numeric: The maximum dimensions of the table}

\item{SType}{character string: The type of plate used, for instance, whether it is control}
}
\description{
Summarize, clean and format the raw data, and store it as a Mongo collection.  Run basic QC before other steps
}
